requirements.txt
# Core LangChain
langchain==0.1.0
langchain-chroma==0.1.4
langchain-huggingface==0.0.3
langchain-ollama==0.1.3
langchain-community==0.0.38

# Vector Database
chromadb==0.4.24
sentence-transformers==2.2.2

# Interface
streamlit==1.29.0
plotly==5.17.0
pandas==2.0.3

# ML/AI
torch==2.0.1
numpy==1.24.4

# Utilities
pyyaml==6.0.1
requests==2.31.0
pathlib2==2.3.7